backend/requirements.txt
fastapi
uvicorn
langchain
chromadb
pydantic
python-dotenv